+++ src/main/java/callprotector/spring/domain/callsession/service/CallSessionServiceImpl.java
@@ -8,15 +8,15 @@ import callprotector.spring.domain.calllog.entity.CallLog;
 import callprotector.spring.domain.calllog.repository.CallLogRepository;
 import callprotector.spring.domain.callsession.entity.CallSession;
 import callprotector.spring.domain.callsession.repository.CallSessionRepository;
+import callprotector.spring.domain.callsttlog.entity.CallSttLog;
+import callprotector.spring.domain.callsttlog.repository.CallSttLogSearchRepository;
 import callprotector.spring.domain.user.entity.User;
 import callprotector.spring.global.apiPayload.code.status.ErrorStatus;
 import callprotector.spring.global.common.enums.CallTrack;
 import callprotector.spring.global.apiPayload.exception.handler.*;
 import callprotector.spring.global.handler.SttWebSocketHandler;
-import callprotector.spring.domain.*;
 import callprotector.spring.domain.abuse.entity.AbuseTypeLog;
-import callprotector.spring.repository.*;
-import callprotector.spring.service.CallSttLogService.CallSttLogService;
+import callprotector.spring.domain.callsttlog.service.CallSttLogService;
 import callprotector.spring.service.GeminiService.GeminiService;
 import callprotector.spring.service.OpenAiService.OpenAiSummaryService;
 import callprotector.spring.domain.user.service.UserService;
