+++ src/main/java/callprotector/spring/domain/callsttlog/dto/response/CallSttLogResponseDTO.java
@@ -1,0 +1,9 @@
+package callprotector.spring.domain.callsttlog.dto.response;
+
+import callprotector.spring.domain.callsttlog.entity.CallSttLog;
+
+public record CallSttLogResponseDTO(
+	String type,
+	CallSttLog payload
+) {
+}
