+++ src/main/java/callprotector/spring/domain/callsttlog/repository/CallSttLogRepository.java
@@ -1,6 +1,6 @@
-package callprotector.spring.repository;
+package callprotector.spring.domain.callsttlog.repository;
 
-import callprotector.spring.domain.CallSttLog;
+import callprotector.spring.domain.callsttlog.entity.CallSttLog;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 import java.util.List;
