+++ src/main/java/callprotector/spring/global/handler/SttContext.java
@@ -12,12 +12,12 @@ import com.google.cloud.speech.v1.*;
 import com.google.protobuf.ByteString;
 
 import callprotector.spring.global.client.FastClient;
-import callprotector.spring.domain.CallSttLog;
+import callprotector.spring.domain.callsttlog.entity.CallSttLog;
 import callprotector.spring.global.common.enums.CallTrack;
 import callprotector.spring.domain.calllog.service.CallLogService;
 import callprotector.spring.domain.callsession.service.CallSessionService;
-import callprotector.spring.service.CallSttLogService.CallSttLogService;
-import callprotector.spring.web.dto.response.CallSttLogResponseDTO;
+import callprotector.spring.domain.callsttlog.service.CallSttLogService;
+import callprotector.spring.domain.callsttlog.dto.response.CallSttLogResponseDTO;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
