--- src/main/java/callprotector/spring/web/dto/response/CallSttLogResponseDTO.java
@@ -1,9 +1,0 @@
-package callprotector.spring.web.dto.response;
-
-import callprotector.spring.domain.CallSttLog;
-
-public record CallSttLogResponseDTO(
-	String type,
-	CallSttLog payload
-) {
-}
